requirements.txt
langchain>=0.1.0
chromadb>=0.4.22
python-dotenv>=1.0.0
openai>=1.12.0
tiktoken>=0.5.2
pypdf>=3.17.1
python-magic>=0.4.27
python-magic-bin>=0.4.14 